refactored: data and forecaster became packages
