+++ src/data/data_extraction.py
@@ -30,21 +30,21 @@ class DataExtraction():
         Extracts the data and saves the row_ids for train, val and test data
         Features will be extracted when a certain row is requested in order to save memory
         """
-        assert p_train+p_val+p_test == 1
+        assert p_train + p_val + p_test == 1
         self.p_train, self.p_val, self.p_test = p_train, p_val, p_test
-        
+
         self.data_dir = dir
 
         # check if files are extracted
         if set(os.listdir(dir)) >= set(["sample_submission.csv", "store.csv", "test.csv", "train.csv"]):
             print("Data is extracted already")
         else:
-            Data.extract(dir+"/data.zip", dir)
+            Data.extract(dir + "/data.zip", dir)
 
         # load into pandas
-        self.store = pd.read_csv(dir+"/store.csv")
-        self.final_test = pd.read_csv(dir+"/test.csv")
-        self.train = pd.read_csv(dir+"/train.csv")
+        self.store = pd.read_csv(dir + "/store.csv")
+        self.final_test = pd.read_csv(dir + "/test.csv")
+        self.train = pd.read_csv(dir + "/train.csv")
 
         self.time_count = self.train.shape[0]
         self.store_count = self.store.shape[0]
@@ -57,23 +57,22 @@ class DataExtraction():
     def test_stuff(self):
         # test if promo2sincewhatever is missing exactly where promo2==constant
 
-        r = {True: [], False: []} # True: where (promo2 is zero)==(value missing)
+        r = {True: [], False: []}  # True: where (promo2 is zero)==(value missing)
         for store_id in range(self.store_count):
             store = self.store.iloc[store_id]
             missing = store["Promo2SinceYear"] is None or store["Promo2SinceWeek"] is None
 
-            r[missing == (store["Promo2"]==0)].append(store_id)
+            r[missing == (store["Promo2"] == 0)].append(store_id)
 
         print("list lemghts", len(r[True]), len(r[False]))
         pdb.set_trace()
-
 
         return
         # are the dates in order?
         dates = [datetime.strptime(d, '%Y-%m-%d') for d in self.train["Date"].tolist()]
         ordered = True
-        for i in range(len(dates)-1):
-            if (dates[i] >= dates[i+1]) != (dates[0] >= dates[1]):
+        for i in range(len(dates) - 1):
+            if (dates[i] >= dates[i + 1]) != (dates[0] >= dates[1]):
                 ordered = False
                 break
         print("Dates are ordered:", ordered)
@@ -113,7 +112,7 @@ class DataExtraction():
             Promo2SinceYear              571 non-null float64
             PromoInterval                571 non-null object
             dtypes: float64(5), int64(2), object(3)
-            
+
         Train
             RangeIndex: 1017209 entries, 0 to 1017208
             Data columns (total 9 columns):
@@ -127,7 +126,7 @@ class DataExtraction():
             StateHoliday     1017209 non-null object
             SchoolHoliday    1017209 non-null int64
             dtypes: int64(7), object(2)
-            
+
         We extract
             Store Type	                One hot 4
             Assortment	                One hot 3
@@ -135,7 +134,7 @@ class DataExtraction():
             CompetitionOpenSinceDays	uint
             PromoSinceDays	            uint if participating in promo, else -1
             DaysSinceInterval	        uint if participating in promo, else -1
-            
+
             DayOfWeek                   One hot 7
             Open                        {0,1}
             Promo                       {0,1}
@@ -154,10 +153,9 @@ class DataExtraction():
             "c": [0, 0, 1]
         }
 
-
         row = self.train.iloc[row_id]
         store_id = row["Store"]
-        store = self.store.iloc[store_id-1]
+        store = self.store.iloc[store_id - 1]
 
         curr_date = self.str_to_date(row["Date"])
 
@@ -168,7 +166,7 @@ class DataExtraction():
         promo_since_days = self._promo_since_days(store, curr_date)
         days_since_interval = self._promo_interval_since_days(store, curr_date) if promo_since_days > 0 else -1
 
-        day_of_week = np.eye(7)[row["DayOfWeek"]-1]
+        day_of_week = np.eye(7)[row["DayOfWeek"] - 1]
         open = row["Open"]
         promo = row["Promo"]
         state_holiday = row["StateHoliday"]
@@ -176,10 +174,11 @@ class DataExtraction():
         state_holiday = abcd[state_holiday]
         school_holiday = row["SchoolHoliday"]
 
-        #features = store_type + assortment + [competition_distance, competition_open, promo_since_days,
+        # features = store_type + assortment + [competition_distance, competition_open, promo_since_days,
         #          days_since_interval] + day_of_week + [open, promo] + state_holiday + [school_holiday]
         features = np.concatenate((store_type, assortment, [competition_distance, competition_open, promo_since_days,
-                  days_since_interval], day_of_week, [open, promo], state_holiday, [school_holiday]))
+                                                            days_since_interval], day_of_week, [open, promo],
+                                   state_holiday, [school_holiday]))
         return features
 
     def str_to_date(self, date_str):
@@ -187,16 +186,16 @@ class DataExtraction():
 
     def _competition_distance(self, value):
         # handles missing values for CompetitionDistance
-        #if value is not None and not np.isnan(value): return value
+        # if value is not None and not np.isnan(value): return value
         if not self._values_missing(value): return value
-        return 100 # TODO
+        return 100  # TODO
 
     def _competition_open(self, store, curr_date):
         # calculates the days since competition is open, handles missing values
         month, year = store["CompetitionOpenSinceMonth"], store["CompetitionOpenSinceYear"]
 
         if self._values_missing(month, year):
-            return 100 # TODO
+            return 100  # TODO
 
         date = datetime.date(day=1, month=int(month), year=int(year))
         return (curr_date - date).days
@@ -209,20 +208,19 @@ class DataExtraction():
             pdb.set_trace()
 
         date = datetime.date(day=1, month=1, year=int(year)) + datetime.timedelta(weeks=int(week))
-        return max((curr_date - date).days, -1) # if at curr_date the store wasn't participating, return -1
+        return max((curr_date - date).days, -1)  # if at curr_date the store wasn't participating, return -1
 
     def _promo_interval_since_days(self, store, curr_date):
         interval_str = store["PromoInterval"]
         starting = interval_str.split(",")
         months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sept", "Oct", "Nov", "Dec"]
-        starting = [months.index(i)+1 for i in starting]
-        started = [datetime.date(day=1, month=i, year=curr_date.year) for i in starting if i<=curr_date.month] + \
-                  [datetime.date(day=1, month=i, year=curr_date.year-1) for i in starting if i>curr_date.month]
-        return min([curr_date-i for i in started]).days
+        starting = [months.index(i) + 1 for i in starting]
+        started = [datetime.date(day=1, month=i, year=curr_date.year) for i in starting if i <= curr_date.month] + \
+                  [datetime.date(day=1, month=i, year=curr_date.year - 1) for i in starting if i > curr_date.month]
+        return min([curr_date - i for i in started]).days
 
     def _values_missing(self, *args):
         return any([a is None for a in args]) or np.any(np.isnan(args))
-
 
     @staticmethod
     def extract(path="data/data.zip", dest="data"):
@@ -230,138 +228,3 @@ class DataExtraction():
         zip_ref = zipfile.ZipFile(path, 'r')
         zip_ref.extractall(dest)
         zip_ref.close()
-
-
-class TimeSeriesData(DataExtraction):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._prepare_time_series()
-
-    def _get_time_series(self, store_id):
-        """
-        :param store_id: store for which the time series will be generated
-        :return date: int number of days since first date of time series
-        :return X: nd.array of shape (batch_size, #features)
-        :return y: nd.array of shape (batch_size, 1)
-        The data returned is only for the specified store, and ordered by date
-        """
-        row_ids = self.train.index[self.train.Store == store_id].tolist()[::-1]
-        days = [self.str_to_date(d) for d in self.train.loc[row_ids]["Date"]]
-        days = [(d-days[0]).days for d in days]
-        X, y = self._extract_rows(row_ids)
-        return days, X, y
-
-    def _prepare_time_series(self):
-        # splits the stores into train, val and test stores
-        train_count = int(self.p_train*self.store_count)
-        val_count = int(self.p_val*self.store_count)
-        self.train_store_ids = set(range(train_count))
-        self.val_store_ids = set(range(train_count, train_count+val_count))
-        self.test_store_ids = set(range(train_count+val_count, self.store_count))
-        self.used_this_epoch = set()
-
-        # create val data
-        pass
-
-        # create test data
-        pass
-
-    def next_train_batch(self, batch_size=50):
-        """
-        :param batch_size: Number of rows, ignored if self.is_time_series is True
-        :return X: nd.array of shape (batch_size, #features)
-        :return y: nd.array of shape (batch_size, 1)
-        Chooses the store and let's _get_time_series do the rest
-        """
-        if len(self.used_this_epoch) == len(self.train_store_ids):
-            self._new_epoch()
-        else:
-            self.is_new_epoch = False
-        store_id = random.sample(self.train_store_ids - self.used_this_epoch, 1)[0]
-        self.used_this_epoch = self.used_this_epoch.union(set([store_id]))
-        return self._get_time_series(store_id)
-
-
-class Data(DataExtraction):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._prepare_random_batches()
-
-    def _prepare_random_batches(self):
-        # splits the rows into train, val and test rows
-        train_count = int(self.p_train * self.time_count)
-        val_count = int(self.p_val * self.time_count)
-        self.train_row_ids = set(range(train_count))
-        self.val_row_ids = set(range(train_count, train_count + val_count))
-        self.test_row_ids = set(range(train_count + val_count, self.time_count))
-        self.used_this_epoch = set()
-
-        # create val data
-        self.X_val, self.y_val = self._extract_rows(self.val_row_ids)
-
-        # create test data
-        #self.X_test, self.y_test = self._extract_rows(self.test_row_ids)
-
-    def next_train_batch(self, batch_size=50):
-        """
-        :param batch_size: Number of rows, ignored if self.is_time_series is True
-        :return X: nd.array of shape (batch_size, #features)
-        :return y: nd.array of shape (batch_size, 1)
-        """
-        if len(self.used_this_epoch) == len(self.train_row_ids):
-            self._new_epoch()
-        else:
-            self.is_new_epoch = False
-        batch_size = min(batch_size, len(self.train_row_ids) - len(self.used_this_epoch))
-        row_ids = random.sample(self.train_row_ids-self.used_this_epoch, batch_size)
-        self.used_this_epoch = self.used_this_epoch.union(set(row_ids))
-        return self._extract_rows(row_ids)
-
-    def get_val_data(self):
-        """
-        :return X: nd.array of shape (val_set_size, #features)
-        :return y: nd.array of shape (batch_size, 1)
-        """
-        return self.X_val, self.y_val
-
-    def get_test_data(self):
-        """
-        :return X: nd.array of shape (test_set_size, #features)
-        :return y: nd.array of shape (batch_size, 1)
-        """
-        return self.X_test, self.y_test
-
-
-class PredictedTimeseriesData(TimeSeriesData):
-    """
-    Get predictions of a forecaster in the same way we get true data
-    Can be used to train a feed forward model and then retrieve time series data for plotting
-    """
-    def __init__(self, true_data, forecaster):
-        """
-
-        :param true_data: Data object
-        :param forecaster: str or AbstractForecaster
-        """
-        if isinstance(forecaster, str):
-            forecaster = AbstractForecaster.load_model(forecaster)
-        self.true_data, self.forecaster = true_data, forecaster
-
-        # modified stuff from super().__init__()
-        self.data_dir = true_data.data_dir
-        self.p_train, self.p_test, self.p_val = 1, 0, 0
-
-        # load into pandas
-        self.store = true_data.store
-        self.final_test = true_data.final_test
-        self.train = true_data.train
-
-        self.time_count = self.train.shape[0]
-        self.store_count = self.store.shape[0]
-
-        self._prepare_time_series()
-
-    def _extract_label(self, row_id):
-        # extracts the sales from the specified row
-        X = self._extract_rows([row_id])
-        return self.forecaster.predict(X)
