+++ src/forecaster/__init__.py
@@ -1,0 +1,4 @@
+from .abstract_forecaster import *
+from .linear_regression import *
+from .feed_forward import *
+from .feed_forward_nn import *
